+++ packages/db/src/types/granule.ts
@@ -4,6 +4,7 @@ export interface PostgresGranuleUniqueColumns {
   collection_cumulus_id: number,
 }
 export interface PostgresGranule extends PostgresGranuleUniqueColumns {
+  producer_granule_id?: string,
   status?: GranuleStatus,
   cmr_link?: string | null,
   error?: object | null,
@@ -32,6 +33,7 @@ export interface PostgresGranule extends PostgresGranuleUniqueColumns {
 // as a "string" type.
 export interface PostgresGranuleRecord extends Omit<PostgresGranule, 'product_volume'> {
   cumulus_id: number,
+  producer_granule_id: string,
   product_volume?: string,
   created_at: Date,
   updated_at: Date,